Took Swagger to the library
