+++ project/apps/customers/src/main.ts
@@ -9,24 +9,19 @@ import { NestFactory } from '@nestjs/core';
 
 import { ConfigService } from '@nestjs/config';
 
-import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
+import { createSwagger } from '@project/lib/swagger';
 
 import { AppModule } from './app/app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
-  const config = new DocumentBuilder()
-    .setTitle('The «Readme» project.')
-    .setDescription('Description of the API server')
-    .setVersion('1.0')
-    .build();
   const globalPrefix = 'api/v1';
-  app.setGlobalPrefix(globalPrefix);
-  const document = SwaggerModule.createDocument(app, config);
-  SwaggerModule.setup('spec', app, document)
   const configService = app.get(ConfigService);
   const port = configService.get<string>('app.port');
   const hostname = configService.get<string>('app.host');
+  createSwagger(app);
+  app.setGlobalPrefix(globalPrefix);
+  app.useGlobalPipes();
 
   await app.listen(port);
   Logger.log(`🚀 Application is running on: http://${hostname}:${port}/${globalPrefix}`);
