+++ project/lib/shared/helpers/src/lib/common.helpers.ts
@@ -1,14 +1,12 @@
 import { ClassTransformOptions, plainToInstance } from 'class-transformer';
 
-type PlainObject = Record<string, unknown>;
-
 const REGEX = /%([^%]*)%/;
 
-export function fillDto<T, U extends PlainObject>(dto: new () => T, plainObject: U, options?: ClassTransformOptions): T;
+export function fillDto<T, P>(dto: new () => T, plainObject: P, options?: ClassTransformOptions): T;
 
-export function fillDto<T, U extends PlainObject[]>(dto: new () => T, plainObject: U, options?: ClassTransformOptions): T[];
+export function fillDto<T, P>(dto: new () => T, plainObject: P, options?: ClassTransformOptions): T[];
 
-export function fillDto<T, U extends PlainObject>(dto: new () => T, plainObject: U, options?: ClassTransformOptions): T | T[] {
+export function fillDto<T, P>(dto: new () => T, plainObject: P, options?: ClassTransformOptions): T | T[] {
   return plainToInstance(dto, plainObject, {excludeExtraneousValues: true, ...options});
 }
