+++ project/lib/swagger/README.md
@@ -1,0 +1,7 @@
+# lib-swagger
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test lib-swagger` to execute the unit tests via [Jest](https://jestjs.io).
